linq rbac模块更新

+++ lind-start-framework/lind-rbac-start/src/main/java/com/lind/rbac/controller/UserController.java
@@ -7,14 +7,18 @@ import com.lind.common.dto.DateRangeDTO;
 import com.lind.common.dto.PageDTO;
 import com.lind.common.rest.CommonResult;
 import com.lind.common.util.CopyUtils;
+import com.lind.logger.anno.LogRecord;
 import com.lind.rbac.dao.PermissionDao;
 import com.lind.rbac.dao.RoleDao;
 import com.lind.rbac.dao.UserDao;
 import com.lind.rbac.dao.UserRoleDao;
 import com.lind.rbac.dto.UserDTO;
+import com.lind.rbac.dto.UserPasswordDTO;
 import com.lind.rbac.entity.Role;
 import com.lind.rbac.entity.User;
 import com.lind.rbac.entity.UserRole;
+import com.lind.rbac.enums.ModuleTypeCons;
+import com.lind.rbac.enums.OperateTypeCons;
 import com.lind.rbac.vo.UserVO;
 import com.lind.redis.service.RedisService;
 import com.lind.uaa.jwt.config.Constants;
@@ -22,13 +26,18 @@ import com.lind.uaa.jwt.config.SecurityUtil;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
+import lombok.SneakyThrows;
+import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.util.CollectionUtils;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
+import javax.validation.Valid;
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Api(tags = "用户接口")
@@ -47,36 +56,45 @@ public class UserController {
     RoleDao roleDao;
     @Autowired
     RedisService redisService;
+    @Value("${defaultPassword:Abcd1234}")
+    String defaultPassword;
     @Autowired
     private PasswordEncoder passwordEncoder;
 
+    private void fillRoleList(UserVO userVO) {
+        List<String> userRoleIds = userRoleDao.selectList(
+                new QueryWrapper<UserRole>().lambda().eq(UserRole::getUserId, userVO.getId())).stream().map(o -> o.getRoleId()).collect(Collectors.toList());
+        if (!CollectionUtils.isEmpty(userRoleIds)) {
+            List<Role> roles = roleDao.selectList(new QueryWrapper<Role>().lambda().in(Role::getId, userRoleIds));
+            if (!CollectionUtils.isEmpty(roles))
+                userVO.setRoleList(roles);
+        }
+    }
+
     @ApiOperation("列表")
-    @GetMapping
+    @GetMapping("list")
     public CommonResult<IPage<UserVO>> index(
+            @ApiParam("用户名") @RequestParam(required = false) String username,
             @ApiParam("时间") DateRangeDTO rangeDTO,
             @ApiParam("页码") PageDTO pageDTO) {
 
         QueryWrapper<User> userQueryWrapper = new QueryWrapper<>();
-        if (rangeDTO.getFromDate() != null) {
-            userQueryWrapper.lambda().ge(User::getCreateTime, rangeDTO.getFromDate());
-        }
-        if (rangeDTO.getToDate() != null) {
-            userQueryWrapper.lambda().le(User::getCreateTime, rangeDTO.getToDate().plusDays(1));
+        Optional.ofNullable(rangeDTO.getFromDate()).ifPresent(o ->
+                userQueryWrapper.lambda().ge(User::getCreateTime, o));
+        Optional.ofNullable(rangeDTO.getToDate()).ifPresent(o ->
+                userQueryWrapper.lambda().le(User::getCreateTime, o.plusDays(1)));
+        if (StringUtils.isNotBlank(username)) {
+            userQueryWrapper.lambda().like(User::getUsername, username);
         }
         userQueryWrapper.lambda().orderByDesc(User::getUpdateTime);
         Page<User> result = userDao.selectPage(
                 new Page<>(pageDTO.getPageNumber(), pageDTO.getPageSize()),
                 userQueryWrapper);
+
         List<UserVO> userVOS = CopyUtils.copyListProperties(result.getRecords(), UserVO.class);
 
         for (UserVO userVO : userVOS) {
-            List<String> userRoleIds = userRoleDao.selectList(
-                    new QueryWrapper<UserRole>().lambda().eq(UserRole::getUserId, userVO.getId())).stream().map(o -> o.getRoleId()).collect(Collectors.toList());
-            if (!CollectionUtils.isEmpty(userRoleIds)) {
-                List<Role> roles = roleDao.selectList(new QueryWrapper<Role>().lambda().in(Role::getId, userRoleIds));
-                if (!CollectionUtils.isEmpty(roles))
-                    userVO.setRoleList(roles);
-            }
+            fillRoleList(userVO);
         }
         IPage<UserVO> userVOIPage = new Page();
         userVOIPage.setRecords(userVOS);
@@ -86,50 +104,75 @@ public class UserController {
         return CommonResult.ok(userVOIPage);
     }
 
-    @GetMapping("{id}")
-    public CommonResult<UserVO> index(@ApiParam("用户id") @PathVariable String id) {
-        return CommonResult.ok(CopyUtils.copyProperties(userDao.selectById(id), UserVO.class));
-    }
-
-
+    @SneakyThrows
     @ApiOperation("新增")
-    @PostMapping
+    @PostMapping("add")
+    @LogRecord(dataId = "${#user.id}",
+            dataTitle = "${#user.username}",
+            moduleType = ModuleTypeCons.USER_MGR,
+            operateType = OperateTypeCons.ADD)
     public CommonResult add(@RequestBody @Validated UserDTO user) {
         if (userDao.selectCount(
                 new QueryWrapper<User>().lambda().eq(User::getUsername, user.getUsername())
                         .or().eq(User::getPhone, user.getPhone())
                         .or().eq(User::getEmail, user.getEmail())) > 0) {
-            return CommonResult.clientFailure(String.format("用户已经存在【%s %s %s】", user.getUsername(), user.getPhone(), user.getEmail()));
+            return CommonResult.clientFailure(String.format("用户已经存在【%s %s %s】",
+                    user.getUsername(),
+                    user.getPhone(),
+                    user.getEmail()));
         }
         user.setPassword(passwordEncoder.encode(user.getPassword()));
         User userEntity = new User();
         CopyUtils.copyProperties(user, userEntity);
         userDao.insert(userEntity);
+        user.setId(userEntity.getId());
         if (!CollectionUtils.isEmpty(user.getRoleIdList())) {
             for (String roleId : user.getRoleIdList()) {
-                userRoleDao.insert(UserRole.builder()
-                        .userId(userEntity.getId())
-                        .roleId(roleId)
-                        .build());
+                UserRole userRole = new UserRole();
+                userRole.setUserId(userEntity.getId());
+                userRole.setRoleId(roleId);
+                userRoleDao.insert(userRole);
             }
         }
         return CommonResult.ok();
     }
 
+    @SneakyThrows
     @ApiOperation("更新")
-    @PutMapping("/{id}")
+    @PutMapping("update/{id}")
+    @LogRecord(dataId = "${#id}",
+            dataTitle = "${#user.username}",
+            moduleType = ModuleTypeCons.USER_MGR,
+            operateType = OperateTypeCons.EDIT)
     public CommonResult update(@ApiParam("用户ID") @PathVariable String id, @RequestBody UserDTO user) {
         User userEntity = userDao.selectById(id);
+        QueryWrapper<User> queryWrapper = new QueryWrapper<User>();
+        queryWrapper.lambda().ne(User::getId, id);
+        queryWrapper.lambda().eq(User::getUsername, user.getUsername())
+                .or().eq(User::getPhone, user.getPhone())
+                .or().eq(User::getEmail, user.getEmail());
+        if (userDao.selectCount(queryWrapper) > 0) {
+            return CommonResult.clientFailure(String.format("用户已经存在【%s %s %s】",
+                    user.getUsername(),
+                    user.getPhone(),
+                    user.getEmail()));
+        }
         if (userEntity != null) {
+
+            Optional.ofNullable(user.getPassword()).ifPresent(o -> {
+                user.setPassword(passwordEncoder.encode(o));
+            });
             CopyUtils.copyProperties(user, userEntity);
             userDao.updateById(userEntity);
+            user.setUsername(userEntity.getUsername());
+
             if (!CollectionUtils.isEmpty(user.getRoleIdList())) {
                 userRoleDao.delete(new QueryWrapper<UserRole>().lambda().eq(UserRole::getUserId, id));
                 for (String roleId : user.getRoleIdList()) {
-                    userRoleDao.insert(UserRole.builder()
-                            .userId(id)
-                            .roleId(roleId)
-                            .build());
+                    UserRole userRole = new UserRole();
+                    userRole.setUserId(id);
+                    userRole.setRoleId(roleId);
+                    userRoleDao.insert(userRole);
                 }
             }
             redisService.del(Constants.USER_PERMISSION.concat(id));
@@ -137,9 +180,56 @@ public class UserController {
         return CommonResult.ok();
     }
 
+    @SneakyThrows
+    @ApiOperation("更新密码")
+    @PutMapping("update-password")
+    public CommonResult resetPassword(@ApiParam("更新密码对象") @Valid @RequestBody UserPasswordDTO user) {
+        User userEntity = userDao.selectById(securityUtil.getCurrUser().getId());
+        if (userEntity != null) {
+            if (!passwordEncoder.matches(user.getPassword(), userEntity.getPassword())) {
+                return CommonResult.clientFailure("老密码不正确");
+            }
+            if (!user.getNewPassword().equals(user.getConfirmPassword())) {
+                return CommonResult.clientFailure("新密码与确认密码不同");
+            }
+            userEntity.setPassword(passwordEncoder.encode(user.getNewPassword()));
+            userDao.updateById(userEntity);
+        }
+        return CommonResult.ok();
+    }
+
+    @SneakyThrows
+    @ApiOperation("重制密码")
+    @PutMapping("bulk-reset-password")
+    @LogRecord(dataId = "${#idList}",
+            moduleType = ModuleTypeCons.USER_MGR,
+            operateType = OperateTypeCons.EDIT)
+    public CommonResult resetPassword(@ApiParam("用户ID列表") @RequestBody List<String> idList) {
+        if (CollectionUtils.isEmpty(idList))
+            return CommonResult.clientFailure("请选择要重置密码的用户");
+        idList.forEach(o -> {
+            User user = new User();
+            user.setId(o);
+            user.setPassword(passwordEncoder.encode(defaultPassword));
+            userDao.updateById(user);
+        });
+        return CommonResult.ok();
+    }
+
+    @ApiOperation("获取用户")
+    @GetMapping("{id}")
+    public CommonResult<UserVO> get(@ApiParam("用户ID") @PathVariable String id) {
+        User user = userDao.selectById(id);
+        if (user == null) {
+            return CommonResult.clientFailure("用户未找到");
+        }
+        UserVO userVO = CopyUtils.copyProperties(user, UserVO.class);
+        fillRoleList(userVO);
+        return CommonResult.ok(userVO);
+    }
 
     @ApiOperation("删除")
-    @DeleteMapping("/{id}")
+    @DeleteMapping("del/{id}")
     public CommonResult del(@ApiParam("用户ID") @PathVariable String id) {
         QueryWrapper<User> userQueryWrapper = new QueryWrapper<>();
         userQueryWrapper.eq("id", id);
@@ -148,18 +238,17 @@ public class UserController {
         return CommonResult.ok();
     }
 
-
     @ApiOperation("批量删除")
     @PostMapping("bulk-del")
-    public CommonResult bulkDel(@ApiParam("用户ID") @RequestParam List<String> ids) {
+    @LogRecord(dataId = "${#ids}",
+            moduleType = ModuleTypeCons.USER_MGR,
+            operateType = OperateTypeCons.DEL)
+    public CommonResult bulkDel(@ApiParam("ID列表") @RequestParam List<String> ids) {
         if (CollectionUtils.isEmpty(ids)) {
             return CommonResult.clientFailure("ids为空");
         }
         for (String id : ids) {
-            QueryWrapper<User> userQueryWrapper = new QueryWrapper<>();
-            userQueryWrapper.eq("id", id);
-            userDao.delete(userQueryWrapper);
-            redisService.del(Constants.USER_PERMISSION.concat(id));
+            del(id);
         }
         return CommonResult.ok();
     }
